修复IME使用时长打点的计算健壮性与会话恢复问题
- ZeroClockMsFromTimeT: 增加mktime失败时返回0的保护
- CalculateDuration: 修复ts比较有符号/无符号转换问题
- Init/OnTimeout: CreateComponents成功后调用RecoverActiveSession恢复活跃会话
- GetNextReportTimeMs: 增加today0溢出保护

+++ services/ime_usage/src/ime_usage_event_cacher.cpp
@@ -352,8 +352,8 @@ void ImeUsageEventCacher::CalculateDuration(
         if (CanCalcDuration(preIt->rawid, it->rawid)) {
             // Use boot time (ts) for inter-event duration — monotonic, immune to
             // wall-clock adjustments (NTP, manual time change).
-            uint64_t duration =
-                (it->ts > static_cast<uint64_t>(preIt->ts)) ? static_cast<uint64_t>(it->ts - preIt->ts) : 0;
+            uint64_t duration = (static_cast<uint64_t>(it->ts) > static_cast<uint64_t>(preIt->ts)) ?
+                static_cast<uint64_t>(it->ts - preIt->ts) : 0;
             // Fallback: screenStatus=0 means uninitialized; treat as UNFOLDED_PORTRAIT(12)
             int32_t status = preIt->screenStatus;
             if (status == SCREEN_STATUS_UNINITIALIZED) {
